fixed task 2 (again)

+++ src/rasterizer.h
@@ -137,7 +137,7 @@ namespace CGL {
     // in preparation for posting pixels to the screen.
     virtual void resolve_to_framebuffer();
 
-    float lineEquation(int x, int y, float x0, float y0, float x1, float y1);
+    float lineEquation(float x, float y, float x0, float y0, float x1, float y1);
     void barycentricCoord(int x, int y, float x0, float y0, float x1, float y1, float x2, float y2, float *coords);
     Color averagePixels(int x, int y);
   };
